Class conversion, optimization, unprocessed texts output, Twitter bootstrap polish, and bug fixes
- TCR has now been converted into a more interface like class for ease
of debugging
- 2-3x speed up with numerous optimizations during each processing step
- Clean up of forked TWiC classes related to MALLET and internal data
structures
- Fixed -1 bug occurring in composite, document-topic weight, and topic
word weight rankings
- All ranks are now more accurate and more accurately reflect the topic
model, as weights with near-nil probability (an assumption originally
based on the most frequently assigned document-topic or topic word
weight) are now utilized instead of a generic small weight.
- Any text file in the corpus which MALLET did not produce any state
file data for are not included in the TCR HTML output. As such, a file
named “tcr_unprocessed.txt” is output in the root directory upon
completion of the script
- CSS added to make the text and rank panels scroll independently